Adding post dorne

+++ domain/model/src/main/java/co/com/musala/model/drone/Drone.java
@@ -15,6 +15,6 @@ public class Drone {
     private final String serialNumber;
     private final String model;
     private final Double weightLimit;
-    private final BigDecimal batteryCapacity;
+    private final Double batteryCapacity;
     private final String state;
 }
